Bux-fixing and converting with async, see modmail.js

+++ src/modules/emoji.js
@@ -8,33 +8,33 @@ class Emoji extends Parse {
     this.find = this.Utility.find;
   }
 
-  react (args) {
+  async react (args) {
     let id = args.shift();
-    this.find([id])
-    .then((msg) => {
+    try {
+      let msg = await this.find([id]);
+      if (!msg) throw "";
       this.emoji(args, msg);
       this.message.delete();
-    })
-    .catch(() => {
+    } catch(e) {
+      try {
       let user = this.Search.users.get(id);
-      if (!user) return this.Output.onError("Couldn't find a user or message matching that ID in this channel.");
-      this.channel.fetchMessages({
-        "limit": 50
-      })
-      .then((_messages) => {
-        let messages = _messages
-          .filter(msg => msg.author.id === user.id && msg.id !== this.message.id)
+      if (!user) throw "Couldn't find a user or message matching that ID in this channel.";
+        let messages = (await this.channel.fetchMessages({
+          "limit": 50
+        })).filter(msg => msg.author.id === user.id && msg.id !== this.message.id)
           .sort((a, b) => {
             return b.createdTimestamp - a.createdTimestamp;
           });
         this.emoji(args, messages.first());
         this.message.delete();
-      })
-    })
+      } catch (e) {
+        if (e) this.Output.onError(e);
+      }
+    }
   }
 
-  emoji (args, message = this.message) {
-    return new Promise ((resolve, reject) => {
+  async emoji (args, message = this.message) {
+    try {
       let succeeded = [], failed = [];
       for (let i = 0; i < args.length; i++) {
         let emoji = this.Search.emojis.get(args[i].trim());
@@ -48,7 +48,7 @@ class Emoji extends Parse {
         }, i * 1000);
         succeeded.push(emoji.id);
       };
-      if (this.command === "gild") return resolve(succeeded);
+      if (this.command === "gild") return succeeded;
       let filter = (reaction, user) => succeeded.includes(reaction.emoji.id) && !user.bot;
       let collector = message.createReactionCollector(filter, {
         "time": 180000
@@ -57,9 +57,11 @@ class Emoji extends Parse {
         reaction.remove(this.client.user)
         .catch(e => console.log(e));
       });
-      resolve(succeeded);
       if (failed[0]) this.Output.onError("Couldn't find emoji " + failed.join(", ") + ".");
-    })
+      return succeeded;
+    } catch (e) {
+      if (e) this.Output.onError(e);
+    }
   }
 
   gild (args) {
